@@ -16,7 +16,7 @@
   ],
   "license": "GPL-3.0",
   "devDependencies": {
-    "eslint": "^8.57.0",
+    "eslint": "^9.0.0",
     "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-prettier": "^5.1.3",
     "eslint-plugin-svelte3": "^4.0.0",
